requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
psutil==6.1.1
sqlparse==0.5.3
tzdata==2025.1
